Do not reset the button state when the menu opens.
Before this patch, if you have hibernate disabled, the hibernate button
shows up transiently whenever you open the system menu, only to disappear
very quickly.
This patch prevents the button dance, but still allows the button to appear
or disappear if the hibernate status has changed (the state is changed
by the callback and the button shows up or hides when the callback
completes).